+++ tests/models/voxtral/test_modeling_voxtral.py
@@ -258,7 +258,7 @@ class VoxtralForConditionalGenerationIntegrationTest(unittest.TestCase):
         ]
 
         model = VoxtralForConditionalGeneration.from_pretrained(
-            self.checkpoint_name, torch_dtype=self.dtype, device_map=torch_device
+            self.checkpoint_name, dtype=self.dtype, device_map=torch_device
         )
 
         inputs = self.processor.apply_chat_template(conversation)
@@ -291,7 +291,7 @@ class VoxtralForConditionalGenerationIntegrationTest(unittest.TestCase):
         ]
 
         model = VoxtralForConditionalGeneration.from_pretrained(
-            self.checkpoint_name, torch_dtype=self.dtype, device_map=torch_device
+            self.checkpoint_name, dtype=self.dtype, device_map=torch_device
         )
 
         inputs = self.processor.apply_chat_template(conversation)
@@ -329,7 +329,7 @@ class VoxtralForConditionalGenerationIntegrationTest(unittest.TestCase):
         ]
 
         model = VoxtralForConditionalGeneration.from_pretrained(
-            self.checkpoint_name, torch_dtype=self.dtype, device_map=torch_device
+            self.checkpoint_name, dtype=self.dtype, device_map=torch_device
         )
 
         inputs = self.processor.apply_chat_template(conversation)
@@ -359,7 +359,7 @@ class VoxtralForConditionalGenerationIntegrationTest(unittest.TestCase):
         ]
 
         model = VoxtralForConditionalGeneration.from_pretrained(
-            self.checkpoint_name, torch_dtype=self.dtype, device_map=torch_device
+            self.checkpoint_name, dtype=self.dtype, device_map=torch_device
         )
 
         inputs = self.processor.apply_chat_template(conversation)
@@ -414,7 +414,7 @@ class VoxtralForConditionalGenerationIntegrationTest(unittest.TestCase):
         ]
 
         model = VoxtralForConditionalGeneration.from_pretrained(
-            self.checkpoint_name, torch_dtype=self.dtype, device_map=torch_device
+            self.checkpoint_name, dtype=self.dtype, device_map=torch_device
         )
 
         inputs = self.processor.apply_chat_template(conversations)
@@ -469,7 +469,7 @@ class VoxtralForConditionalGenerationIntegrationTest(unittest.TestCase):
         ]
 
         model = VoxtralForConditionalGeneration.from_pretrained(
-            self.checkpoint_name, torch_dtype=self.dtype, device_map=torch_device
+            self.checkpoint_name, dtype=self.dtype, device_map=torch_device
         )
 
         inputs = self.processor.apply_chat_template(conversations)
@@ -491,7 +491,7 @@ class VoxtralForConditionalGenerationIntegrationTest(unittest.TestCase):
         disclaimer: Perfect token matching cannot be achieved due to floating-point arithmetic differences between vLLM and Transformers implementations.
         """
         model = VoxtralForConditionalGeneration.from_pretrained(
-            self.checkpoint_name, torch_dtype=self.dtype, device_map=torch_device
+            self.checkpoint_name, dtype=self.dtype, device_map=torch_device
         )
         inputs = self.processor.apply_transcription_request(
             language="en",
